backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
alembic==1.14.1
python-multipart==0.0.20
python-dotenv==1.0.1
pytesseract==0.3.13
Pillow==11.1.0
pypdf==5.1.0
pdf2image==1.17.0
langchain==0.3.15
langchain-community==0.3.15
langchain-ollama==0.2.3
chromadb==0.5.23
reportlab==4.2.5
httpx==0.28.1
pytest==8.3.4
